Update README.md
The main purpose of the application is to allow users to upload an 
image and analyze its contents using Google Cloud Vision API. The 
API can detect labels (objects), text (OCR), faces, landmarks, logos, 
and more. The application presents this information to the user in a 
simple, user-friendly interface built with CustomTkinter.